Move stats helpers into stats endpoint
